TINKERPOP-2534 Switched test/distribution logging to logback
There is a CVE with log4j 1.2 which EOL'd 5 years ago. We haven't really focused on changing this earlier because log4j is an optional dependency and can be swapped out by the user to their preferred logging provider given that we use slf4j. That said, it's better that we simply not include it in our distributions and docker packaging would work better out of the box without the CVE struck log4j in the mix. logback is dual licensed, but given that one of those licenses is under EPL we can make use of it - see https://issues.apache.org/jira/browse/LEGAL-63 for futher details if needed.