TINKERPOP-2554 Initialized metrics on ProfileStep for getMetrics()
Only happens if the traversal is in a locked state, otherwise you get a null. For 3.6.0 we can change the signature of to return an Optional to better support this eventuality. Did not change DefaultTraversalMetrics behavior where the NPE was thrown since initializing in getMetrics() should prevent this error. CTR