TINKERPOP-2585 Adjust strategy order for InlineFilterStrategy
The test associated with TINKERPOP-2585 in the JIRA wasn't easy to translate to something generalized for our test suite and what's more it was somewhat non-deterministic. Where it was failing for me consistently some time ago, as of today, I wasn't able to get it to fail and I dont think anything else changed on the way to it being fixed. When it was failing this extra bit of order definitely fixed the problem.