Pin peer dependency packaging to 20.9
The Python build is currently broken which is apparently caused by
this peer dependency in version 21.
Pinning it to 20.9 seems to fix the build.
We can remove this pinning later when the problem is gone (so the
builds goes through without this pinning) or when we drop 3.4-dev
as the problem seems to be limited to that branch (probably Python
2).