TINKERPOP-2585 Strategies were being applied when unecessary
Not sure if this was related to the original issue but decided to tack this in here, but basically strategies were being repeatedly applied under the 3.5.0 strategy model when they shouldn't have and were therefore yielding incorrect results at times. Hard to say exactly what might have been affected but the issues surfaced in the new test that was added here.