TINKERPOP-2601 Neo4jGraph running gherkin tests
Includes full build over GLVs after a number of changes that were needed in feature files to better support ordering of results and other odds and ends that Neo4j needed for the tests to pass. Added annotations for multi/metaproperty tests so that Neo4j and other graphs could ignore those.