Removed optional "integrate" and "performance" test suites
Performance tests were removed long ago and there was only one integrate test which really just validated serialization size which really should be a given (though at the time it wasn't). The time on that test ran long and i think it just moved to "integrate" tests because it was "long" but a pool of integration tests never really developed. CTR