TINKERPOP-2646 Make StreamExtensions accessible
They can now also be used by users and driver providers who implement
their own custom GraphBinary types.
I decided to change the signature of `ReadAsync()` a bit as it's easier
to use if only the number of bytes to read has to be provided instead
of a buffer that needs to have the correct length.