TINKERPOP-2635 Improved consistency of by() behaviors
Added ProductiveByStrategy to improve consistency of by() and to establish a compatibility bridge for 3.6.0 to have 3.5.x behavior. The strategy is non-breaking in that the only behavior it removes is exception behavior which will be unlikely to break any code as no one is likely writing Gremlin designed to fail.