Refaced website
No real changes to content - it has mostly just been re-arranged. There is no more header-footer.html or templating so if the the header/footer changes it must be changed everywhere. Given how things have restructured this isn't too burdensome and those things tend not to change all that much.