TINKERPOP-2682 Enable WS compression in .NET
WebSocket support was only added in .NET 6 which made it necessary to
explicitly add .NET 6 as an additional target framework.

I also enabled package validation [1] so we a) notice if we introduce
breaking changes by accident and b) ensure that our code compiled
against .NET Standard 2.0 can also be run against .NET 6.
This already helped me to keep the breaking change here to a minimum.
It's just an added optional parameter.
We could refactor the GremlinClient constructors in the future in a way
to avoid this problem of having to add optional parameters for new
config option which is always a breaking change.
The validation is executed on `dotnet pack` which gets already executed
via `mvn verify` and is therefore also part of our GH actions.

Compatibility errors that we know about and that we accept should be
added to the `CompatibilitySuppressions.xml` file which can also be
generated via

```
dotnet pack Gremlin.Net.csproj /p:GenerateCompatibilitySuppressionFile=true
```

WebSocket compression can be problematic due to attacks like
CRIME/BREACH as noted in TINKERPOP-2682. So, I added an option to easily
disable it for applications that might be vulnerable and added docs to
make users aware of this. We might want to do the same maybe also for
Java and Python.

Lastly, some quick benchmarks I ran on my machine:

GraphBinary:

Without compression:
Packets transmitted: 3594
Bytes transmitted: 7MB
Runtime: 1min 1.5sec

With compression:
Packets transmitted: 62 (1.7%)
Bytes transmitted: 34KB (0.5%)
Runtime: 1.3sec

GraphSON 3:

Without compression:
Packets transmitted: 12029
Bytes transmitted: 25MB
Runtime: 1min 40sec

With compression:
Packets transmitted: 116 (1.0%)
Bytes transmitted: 103KB (0.4%)
Runtime: 1min 39sec

So, both for GraphSON as well as for GraphBinary the network traffic
shrinks by a factor of ~100 while the total runtime is not affected.

[1]: https://docs.microsoft.com/en-us/dotnet/fundamentals/package-validation/overview