TINKERPOP-2518 Enable more Gherkin scenarios
There were three categories of scenarios ignored that are now enabled:
* Some scenarios passed without any change. I don't know why they were
ignored in the first place. Maybe we've fixed something in the meantime
but forgot to enable them again.
* Some scenarios expect a decimal type as the result but the server
sends a double because .NET also uses double by default in its
traversals. So I changed the assertion to pass if the values are equal
although the types don't match (for decimal and double only).
* One scenario required a more complex assertion as it expects an array
of dictionaries which means that we have to perform a deep equality
check.