TINKERPOP-2518 Fix remaining scenarios on master
This uncovered a bug in the HasId step that threw a
`NullReferenceException` if the second argument was `null`. We probably
have the same problem with other steps that are just not tested yet for
`null`. I however haven't touched them here as I'm not sure yet where
we actually want to allow `null`.

Ignoring the tag `AllowNullPropertyValues` automatically in .NET simply
didn't work as we checked for the tag on the feature, but it's on the
scenario.