TINKERPOP-2705 Improve handling of null arguments
`GraphTraversal`/`GraphTraversalSource`: Ensure that no
`NullReferenceException` gets thrown. Either allow `null` as an argument
or throw an `ArgumentNullException` if it doesn't seem sensible and if
Java also doesn't allow `null`.
__: Simply forward all `null` arguments to the steps in
`GraphTraversal` so they can be handled there and we don't need to
duplicate the logic whether we want to allow `null` or throw.

I had to adapt the `DotNetTranslator` a bit as the logic to avoid
problems where multiple overloads would be possible was to eager. This
resulted in a `params object[]` argument that should be set to `null`
for a scenario, being actually tested with an object[] which had `null`
as an element. This let the scenarios pass although users would get a
`NullReferenceException` since they probably won't add the unnecessary
`object` cast.
Example:
`HasKey('name', null)` -> threw a `NullReferenceException`
`HasKey('name', (object) null)` -> passed