Update board reports docs
The Apache Reporter Service now seems to be the preferred way to
submit board reports, but Whimsy is also a good alternative.

I've removed the template for the report because I've just changed the
format in the last report according to what's suggested by the reporter
service so we now have two different formats that we've used.
I also think that anybody who might at any point submit such a report
has already read enough other reports so that they know the format and
have enough older reports as a reference.