TINKERPOP-2742 Give graph providers a hint of property cardinality
Currently, Graph::getCardinality only takes a vertex property key as the parameter,
which sometimes is limited for the graph provider to make a proper choice
about the cardinality of the vertex property. This commit overloads this method
which takes both the property key and values as parameters so that the graph
provider can hopefully make a better choice. For example, a graph provider can
use list/set cardinality when it sees multiple values and uses single
cardinality when it only sees one value. This is particularly useful when loading
a graph using the IO step without a pre-defined schema.