Fixed build issues in recently merged GraphBinary for javascript
Build issues only occurred after merge to 3.6-dev where new tests were added and mergeE/V() tests were present that could not have been accounted for during development on 3.5-dev. Also had to fix an issue with the construction of VertexProgramStrategy which was shipping nulls to JavaTranslator that it could not handle. CTR