gremlin-go: support per-request arguments in bytecode
The issue that triggered this change is that per-request options were not
working with gremlin-go. For instance, when a per-request timeout was set this
way:

	result, err := g.With("evaluationTimeout", 5000).V().Count().Next()
	if err != nil {
		// Handle error.
	}

A very similar problem was described some time ago for the Python GLV in the
JIRA issue TINKERPOP-2296 [1] and solved in the pull-request
apache/tinkerpop#1329 [2].

Stephen Mallette explains the root cause of the issue in a comment in JIRA:

	There is no real strategy application in GLVs - the strategies are
	simply shipped to the server as bytecode configurations. By that point,
	it's already bypassed the point where the server will care about the
	timeout (or other request setting) overrides.

The Python GLV fixes this issue by:

1. Converting the "with" options into an OptionsStrategy.
2. Extracting a whitelisted set of options from the OptionsStrategy.
3. Sending the extracted options as part of the bytecode request.

However, this means that the generated bytecode will not always match the code
written by the users. E.g.

	with(k, v) -> withStrategies(OptionsStrategy({k: v}))

So, this PR takes a slightly different approach. It extracts the whitelisted
set of per-request options in makeBytecodeRequests() and adds them to the
bytecode request without modifying the generated bytecode.

[1] https://issues.apache.org/jira/browse/TINKERPOP-2296
[2] https://github.com/apache/tinkerpop/pull/1329