gremlin-go: use docker compose file format version 3.7
The current docker-compose.yml file is not compatible with the docker-compose
version packaged by some widely used Linux distributions. For instance, Debian
bullseye (current stable) packages docker-compose 1.25.0 [1].

docker-compose 1.25.0 supports up to compose file format 3.7 [2] and requires
the version to be explicitly set in the docker-compose.yml file.

By setting the version of the docker-compose.yml file to 3.7 we support
docker-compose 1.22.0+ and the only required change is adding the "version"
field. Also, version 3 has not been deprecated [3].

[1] https://packages.debian.org/bullseye/docker-compose
[2] https://github.com/docker/compose/releases/tag/1.25.0
[3] https://docs.docker.com/compose/compose-file/compose-versioning/