Update docker/build.sh to work with GLV docker-compose changes.
The non-Java GLVs have changed to use docker-compose for integration
testing which creates a Docker in Docker issue when run through the
build.sh script. To prevent this issue, the GLVs are split away and run
individually using their docker-compose files. However, since the
docker-compose files require some artifacts from the build (including
the local Maven repository items), those directories are bind mounted so
that the files are available for the docker-compose builds.

To increase the usability of the script for developers who only work on
a portion of TinkerPop, options were added to include only a subset of
the modules to the build.

Change version parsing to use awk instead of grep with Perl regex
because it isn't installed by default on macOS.