TINKERPOP-2794 Add support for cancellation
This enables users to cancel the evaluation of a single traversal /
the execution of a submitted script.
Cancellation on the server is however out of scope for this contribution
as we first need TINKERPOP-2210 for that. This means that we only cancel
client-side stuff here, like if a request was not (completely) submitted
yet to the server, then we can still cancel it. If it was already
submitted, then we just stop waiting for the response from the server.
This is of course not ideal, but it at least frees up resources for the
client.

If the user however wants to shut down the application completely, then
they should simply Dispose() the `IGremlinClient` /
`DriverRemoteConnection`.

How we handle cancellation in the different components:

Connection
    WebSocketConnection
        ConnectAsync called before any requests come in
            -> leave unchanged
        ReceiveAsync called in a loop
            -> don't cancel as as it can still be used for other
            requests
        SendAsync called for this request
            -> pass CancellationToken
        CloseAsync called if the ConnectionPool gets disposed or if the
            connection got closed, e.g., by the server or due to a
            network problem
            -> no point in cancelling this
    multiple ReceiveHandlers
        (`IResponseHandlerForSingleRequestMessage`), each:
            Task the user is awaiting on, corresponding to a single
            request
                -> needs to be canceled if cancellation is requested
        The handler for the cancelled request can be removed afterwards.
    IMessageSerializer
        SerializeAsync / DeserializeAsync -> pass CancellationToken

IMessageSerializer
    GraphSONMessageSerializer
        operations are sync -> we can only check if cancellation is
            already requested before we call them
    GraphBinaryMessageSerializer
        operations are async & support cancellation
            -> pass CancellationToken

We unfortunately cannot cancel the deserialization of received messages
as the CancellationToken is bound to the specific request which it can
only identify based on the RequestID, but that needs to be deserialized
first.
That is also why we still deserialize responses for requests that were
already cancelled.