Cole/m1 support (#38)
Compatibility fixes for ARM based Macs

Updated base images for gremlin-server and gremlin-console docker images to support arm64:

The previously used adopt openjdk11 alpine linux image does not support arm64.
In fact none of the major openjdk distributions maintain a jdk11 alpine image for arm64.
Changed base image to alpine and installed oracle openjdk 11. Oracle openjdk 11 was chosen as it is already maintained in the APK community repository.

Added jansi as an explicit dependency of gremlin-console. Previously it was a dependency of jline which itself is a dependency of groovy. The console is able to run just fine on x64 architectures but it cannot run on arm JDK's without importing Jansi directly.

Updated dev docs with instructions to configure docker desktop for better performance on apple silicon.