TINKERPOP-2817 Support java.lang.Byte in hadoop Graphson
Hadoop GraphSONRecordReader and GraphSONRecordWriter does not support
'java.lang.Byte' type, which is supported by both GraphSONXModuleV2d0 and
GraphSONXModuleV3d0, but both of them are custom modules and excluded by default.

This fix adds GraphSONXModuleV2d0 and GraphSONXModuleV3d0 as custom
module for GraphSON V2_0 and V3_0 by default to avoid the failure of
TINKERPOP-2817.