Refactoring java driver and extracting SimpleSocketServer to new module
This commit represents a significant restructuring of gremlin-driver. The intention behind these changes is to make SimpleSocketServer usable by all of the GLV's and work towards bringing the GLV's up to the same testing standards as the java driver. This commit aims to do so without introducing any breaking changes such that the benefits of improved testing can be realized sooner. Further breaking changes may be in order in some release where it is permittable. The changes are detailed below:

Moved Simple-Socket-Server and related initializer classes from gremlin-driver into a new module gremlin-tools/gremlin-socket-server. This is so other GLV's can depend on the socket server for testing without depending on the whole gremlin-driver. In this change the classes are only moved, the package remains tinkerpop.gremlin.driver to avoid breaking any code which uses these classes.

A new gremlin-util module is created to house any code which is to be shared between gremlin-driver, gremlin-server, and gremlin-socket-server. Previously all of these shared classes (such as serializers and RequestMessage...) were part of gremlin-driver (which is a dependency of gremlin-server). These classes needed to be extracted to a new module to prevent a cyclic dependency between gremlin-driver and gremlin-socket-server. This structure also removes the need for gremlin-server to depend on gremlin-driver although this dependency should remain for the time being to avoid breaking any users who may indirectly use gremlin-driver classes through the server.

Builds SimpleSocketServer into a docker image and moves hardcoded constants for the server port and custom request id's into a config yaml. These changes bring gremlin-socket-server into a state which can be utilized by all of the GLV's as opposed to only the java driver.

Modified TestWSGremlinInitializer to serialize single vertices as a list to more closely emulate GremlinServer.

Notes Before Merging this commit:

If merging into 3.5.x, ensure that the deprecated gryo serializers are moved to gremlin-util alongside the other serializers. No changes to these files are necessary.