TINKERPOP-2348 Add nullable annotations
This only adds nullable annotations to the projects Gremlin.Net and the
Gremlin.Net.Template, but not to the test projects.
I intend to add it there in a follow-up PR to at least somehow limit
the scope of this commit. This takes already care of everything
user-facing.

This unfortunately includes a few breaking changes (not necessarily
complete):
- Message types:
  * Some set property accessors removed
- Structure types:
  * Some set property accessors removed
- DefaultTraversal.Bytecode now abstract
- GraphBinaryReader/Writer
  * ReadValueAsync / WriteValueAsync were split into two methods,
        for nullable and non-nullable values
- Changed exceptions: Instead of NullReferenceExceptions,
    IOExceptions should be thrown during serialization

Most of these breaking changes are also mentioned in the upgrade docs.
I haven't added all of them there as I found it to be very unlikely that
anyone is really affected by them.