TINKERPOP-2348 Add nullable annotations also to tests
Only the generated Gherkin test traversals are missing as we would have
to let the Java `DotNetTranslator` produce C# code with nullable
annotations. I think we should do that at some point in the future as
nullable annotations are only enabled since .NET 6 by default and only
for new projects. So, many users are probably still not using them.

Adding the annotations also in tests revealed two (well actually three)
more places where `null` should be allowed in Gremlin.Net itself:
In `Traversal.Equals()` and in the `From()`/`To()` steps if the argument
is a `Vertex`. While `null` is still not really allowed there, Gremlin
throws a more meaningful exception message if it gets `null` and we
might change that in the future to somehow allow `null` maybe.