TINKERPOP-2804 Moved gherkin files to gremlin-test resources
Taking this approach makes it easier for providers to reference the files from the classpath, otherwise they likely need to have copies of them on the local file system which isn't so good. This is a breaking change but it only affects providers. They will have a minor adjustment to make to their test infrastructure on upgrade.