Add User-Agent connection header to Javascript GLV
User agent follows the form of [Application Name] [GLV Name].[Version] [Language Runtime Version] [OS].[Version] [CPU Architecture]

New feature in the javascript driver to send a user agent header during the web socket handshake for new connections.
This behavior is enabled by default but can be disabled with a new enableUserAgentOnConnect option.

This commit also adds a configuration section to the JS reference documentation in order to be consistent with the other GLV's.
The configuration table is populated based on the JSDoc comments from the Connection, Client, and DriverRemoteConnection constructors.