Fix RequestId Override CTR
Fixes a bug where python and go glv's were not using the correct request_id
when overriden. This is detailed in TINKERPOP-2841 and TINKERPOP-2844. This
commit is not sufficient to close those tickets as further investigation is
needed to ensure the other per-request settings work as intended, this
commit provides an immediate fix to the broken request_id functionality.