Improved handling of cases where Gremlin Version is not available
Gremlin.java:
Changes Gremlin.version() such that it returns "VersionNotFound" in any
case where reading the version from the manifest fails. This is a better
result than the previous case where a ExceptionInInitializerError was thrown.

For java user agent:
Wraps all elements of UserAgent in a try catch and default's to a value of
"NotAvailable" if an element fails.

For JS user agent:
Added error handling in case client cannot read from package.json