Fix server not handling StackOverflowError in bytecode traversals (#1979)
Fixes TINKERPOP-2767.

The existing error handling in TraversalOpProcessor and SessionOpProcessor during bytecode iteration was
catching all exceptions sending an appropriate error response to the client. The OpProcessor however was
not catching a StackOverflowError which could be induced by running a query which contains a
large repeat step. This Error was being caught by FutureTask.run() but GremlinServer never wait's
on this task or checks the results which caused this Error to be lost.
A similar issue was found to exist in AbstractSession.process().

This commit adjusts the existing error handling code to catch any Throwable during bytecode
iteration so clients will receive error messages and codes for server errors as well as exceptions.
Any errors are re-thrown such that the evalFuture FutureTask will continue to have an exception set
correctly (although GremlinServer currently does not use this for anything).