TINKERPOP-2373 Bump to Groovy 4.0.9/GMavenPlus 2.1.0
Groovy 4 packages the indy version by default so it isn't necessary to
use the classifier anymore. asm-util is a new dependency for groovysh
but clashes with another asm dependency so we exclude it from groovysh
since it isn't required when running (Groovy will use the version of asm
from jarjarasm). Removed usages of "def class" since this no longer
seems to be supported by Groovy. Updated the parameter size for
shouldProvideBetterExceptionForMethodCodeTooLarge test as it no longer
triggers the exception with the previous, smaller size.

Co-authored-by: Stephen Mallette <spmva@genoprime.com>