TINKERPOP-2899 Changed hashCode implementation for LP/P traversers.
In some instances, a graph may contain elements whose Ids are
sequentially increasing integers. Because the hashCode for Path
implementations are calculated based on these same Ids, you run into a
situation where both parts of the hashCode calculation are related by
some constant. H(A)-H(B)=C where C is a constant. In this case, XOR is a
bad candidate for creating the final hashCode because XOR will reduce to
some small pool of potential output. This occurs because the most
significant bits will be the same for H(A) and H(B) which will cause
them to become 0 when XOR'd. This commit changes the algorithm to be a
more standard algorithm for this case which is m * H(A) + H(B).