TINKERPOP-2957 allow cardinality to be specified for mergeV()
Provides a nicer way to allow users to control cardinality of property values when using mergeV() and property(Map). A bit tough to get this in without changes to the serializers but by treating the CardinalityValue as a special form of Bytecode it allowed this change to keep the serializers sacred.