TINKERPOP-2948 Add StreamReadConstraints config to GraphSON serializers
With the upgrade to jackson 2.15.0, new StreamReadConstraints are
introduced which set certain max token sizes for deserialization.

This commit is adding configuration options to all GraphSON serializers
to define maxNumberLength, maxStringLength, and maxNestingDepth options.

Also includes a small change to the shade plugin to stop generating
dependency-reduced-pom.xml's as they were leading to dependency
related build problems throughout the project.