TINKERPOP-2976 Fix modification while iterating
In rare cases users could get an InvalidOperationException if the
dictionary of `_serializerByType` was modified while another thread in
parallel iterated over the keys of this dictionary.
Just creating a copy of the supported types and then iterating over that
should fix this.
The problem was reported for GraphBinary, but the GraphSONWriter had the
same problem so I also fixed it there.