TINKERPOP-2724 Enabled some more gherkin tests around Path
Path results have not been asserting for Java due to path labels not being defined in gherkin. As a result Path.equals() was always returning false. There was an assumption in the Java StepDefinition that simply ignored those tests that had that type of assertion so they simply werent running. This enabled about 30 additional tests. CTR