CTR Fix Python User Agent Test
Recently some extra alphanumeric chars started to be appended to the end
of the version in the default aiohttp user agent which conflicted with the
rigid regex pattern the test was looking for.

This change relaxes the pattern as this test is only intended to show that
the gremlin user agent is not being sent.