Update type parameters for reverse and merge CTR.
These two steps can return different types depending on the incoming
traverser. The original design used a wildcard to deal with this but
this could be frustating for users as they would always need to cast.
This update essentially changes the type to Object to reduce the
potential amount of casting the user needs to do.