Loosen error message validation a bit by ignoring case.
Not sure this is a good long term solution but perhaps it will help a bit. We've been a bit oppressive with error message assertions in the past and it hasn't really worked well. Asserting just an error doesn't seem right either without some qualification that it is the right one. CTR