TINKERPOP-3056 Fixed bug in mid-traversal mergeE for onMatch
Since the current traverser was only being updated onMatch when it was used as a start step doing mutations in a sideEffect() would update whatever the current traverser was. This code may have been leftover from a time when the merge steps operated more readily on the current traverser as an argument to the step.