TINKERPOP-3063 Fix authentication for concurrency
Executing multiple queries in parallel could lead to authentication
failures if `MaxInProcessPerConnection` is set to a value higher than
`1` as the second request could then be sent to the server while the
server was still waiting for the authentication challenge response from
the driver for the first query.

This made it necessary to set `MaxInProcessPerConnection=1` as a
workaround for such scenarios which of course means that connection
pooling is less efficient.

Simply sending a validation request on each connection right after
establishing the connection and therefore before it can be used to
submit queries from the user should fix this issue.
The downside of this is of course that connection establishment takes
slightly longer. I think this is an acceptable trade-off even for
scenarios where authentication is not used since this also validates the
connection irrespective of authentication and also because this delay
only occurs once right at the start of an application.