Allow aliased client to pass along options (#2524)
Refactored DriverRemoteConnection.getRequestOptions() to be a public static method in RequestOptions so that Client.submitAsync() can get and pass along options with it. This fixes the issue where aliased clients submitting traversals containing with() configurations would ignore sending these options as part of the request in https://issues.apache.org/jira/browse/TINKERPOP-3039.