Updated tests that depended on PARTIAL_CONTENT status CTR
These tests needed updating because the PARTIAL_CONTENT status is not
used for HTTP and was from the previous WebSocket subprotocol. For the
newer HTTP handlers, there isn't a way to signal which ResponseMessage
is the one that contains the actual error due to the addition of
trailers at the end. So, the conditionals are slightly changed to look
for error in any response.