Custom Strategy Construction (#2806)
This commit adapts TraversalStrategyProxy in Java to work based off of strategy names instead of strategy classes. This change allows the proxy to be used in traversals in order to pass custom/provider defined strategies and configurations into the gremlin script. The advantage of this approach is that the user does not need any concrete strategy class for custom strategies, and the only actions needed from a provider to support a custom strategy is to either register it for use with gremlin-lang, or add it to an import customizer for gremlin-groovy. The TraversalStrategy base class in python is similarly adapted to pass custom strategies into gremlin scripts.