Fix NullPointerException being thrown by outbound encoder. (#2800)
The NPE was due to null remote host which can happen with TLS 1.3 client
certification auth failure. Client cert auth failure can happen after a
successful handshake, causing an inbound SSLException. Changed the inbound
handler to detect and store the SSLException in the channel attributes and
changed the outbound encoder to detect if remote address is not available
and there was an SSLException. Changed the integration test to log the
stack trace of the caught exception for easier troubleshooting as these
SSL tests have been flaky.