Update RequestInterceptor to better work with HTTP. (#2799)
The interceptor is updated to run for every request to allow modifying
each HTTP request. It works on a newly created abstraction of a request
called HttpRequest to prevent breaking changes in the event the underlying
library ever changes. A name is required when adding/removing
interceptors as it's the simplest way to identify interceptors that were
inserted from lambdas.