Added request interceptor to gremlin-python.
The message_serializer is split into the request_serializer and the
response_serializer to allow for setting the request_serializer to None
which prevents payload serialization. This makes it easier for the user
to modify the payload. Auth is not combined with interceptor as it
requires refactoring to provide the URL in protocol and simplifies usage
compared to the Java driver albeit at the cost of flexibility.