Add log entry in WsAndHttpChannelizerHandler
The log entry here is necessary because when an exception thrown by
Netty reaches exceptionCaught(), if the error is not printed, the
server side will not be able to know what has happened.

For example, if the maxContentLength in JanusGraph's Gremlin-Server
configuration is set to 65536, and the client sends data over this
size via WebSocket, the JanusGraph server logs will appear normal,
but the client will experience a disconnection, making it difficult
to immediately identify the cause of the issue.