WIP Remove unused "Arguments" from grammar.
For any steps which do not accept a GValue in GraphTraversal,
this commit updates the grammar to reduce the aruments from
"___Argument" to "___Literal"

At this state, the changes are focused towards String arguments
and literals. This work needs to be continued to all argument
types.

Note that currently, the grammar uses a naming convention for
varargs such that "stringLiteralVarargs" is defined as 0 or more
"StringNullableArgument", not "StringNullableLiteral". This
naming convention for varargs is misleading, although I have not
addressed it in this commit to avoid unnecessary clutter to the
diff. This commit introduces a new type "stringLiteralVarargsLiterals"
as a temporary placeholder for varargs of actual "StringNullableLiteral".

After this work has been reviewed, I intend to followup with a CTR
which renames stringLiteralVarargs to stringArgumentVarargs, and then
the new stringLiteralVarargsLiterals can assume the correct name of
stringLiteralVarargs.