TINKERPOP-2959 Restrict Arguments in Grammar
The change from DirectVariableResolver (always produces literals) to the new
DefaultVariableResolver (produces GValues) leads to unintended consequences
if users attempt to pass a variable into a step which does not accept GValue.

The grammar is being reduced to only accept variables in places which parameterization
is explicitly intended, in steps which accept GValue.

This commit additionally adds many more tests to ensure appropriate coverage for all
parameterizable steps.

Serializers for the Merge enum have been reintroduced in order to allow parameterization
of the maps for mergeE().

Note that currently, the grammar uses a naming convention for
varargs such that "stringLiteralVarargs" is defined as 0 or more
"StringNullableArgument", not "StringNullableLiteral". This
naming convention for varargs is misleading, although I have not
addressed it in this commit to avoid unnecessary clutter to the
diff. This commit introduces a new type "stringLiteralVarargsLiterals"
as a temporary placeholder for varargs of actual "StringNullableLiteral".

This work is intended to be followed by a CTR
which renames stringLiteralVarargs to stringArgumentVarargs, and then
the new stringLiteralVarargsLiterals can assume the correct name of
stringLiteralVarargs.

Co-authored-by: Yang Xia <55853655+xiazcy@users.noreply.github.com>
Co-authored-by: Ken Hu <106191785+kenhuuu@users.noreply.github.com>