Promote gremlin-lang as default over gremlin-groovy (#2966)
Makes gremlin-lang the implicit (default) language, replacing gremlin-groovy, for any request which does not explicitly specify a language. gremlin-groovy is left intact for any requests which explicitly set as its language.

This change is motivated by the removal of bytecode and the new HTTP API in TP4. This change makes the server more aligned with the new script-based GLVs which produce gremlin-lang.